# optimize serverside smoke logic