even if there were no NPC's nor smokes on the map it drained a lot of performance; now it requires both to be present to do visibility checks

before:
no NPCs present; ~0.68ms/tick
5 NPCs present, no smokes; ~0.9ms/tick

after:
no NPCs present; <0.003ms/tick
5NPCs present; no smoke; <0.003ms/tick
5 NPCs present; smokes; ~0.13ms/tick